+++ src/cli/index.ts
@@ -1,5 +1,6 @@
 #!/usr/bin/env bun
 import { Command } from "commander";
+import { registerEventsCommands } from "@hasna/events/commander";
 import { runCommand } from "./commands/run.js";
 import { ciCommand } from "./commands/ci.js";
 import { judgeCommand } from "./commands/judge.js";
@@ -34,5 +35,7 @@ program.addCommand(mcpCommand());
 program.addCommand(captureCommand());
 program.addCommand(completionCommand());
 program.addCommand(syncCommand());
+registerEventsCommands(program, { source: "evals" });
+
 
 program.parse(process.argv);